+++ Cinnamenu@json/files/Cinnamenu@json/utils.js
@@ -1,6 +1,6 @@
 const Mainloop = imports.mainloop;
 
-var setTimeout = function(func, ms) {
+const setTimeout = function(func, ms) {
   let args = [];
   if (arguments.length > 2) {
     args = args.slice.call(arguments, 2);
@@ -14,11 +14,11 @@ var setTimeout = function(func, ms) {
   return id;
 };
 
-var clearTimeout = function(id) {
+const clearTimeout = function(id) {
   Mainloop.source_remove(id);
 };
 
-var setInterval = function(func, ms) {
+const setInterval = function(func, ms) {
   let args = [];
   if (arguments.length > 2) {
     args = args.slice.call(arguments, 2);
@@ -32,21 +32,15 @@ var setInterval = function(func, ms) {
   return id;
 };
 
-var clearInterval = function(id) {
+const clearInterval = function(id) {
   Mainloop.source_remove(id);
 };
 // Native objects such as CinnamonApps and MetaWindows stringify with a unique identifier.
-var isEqual = function(a, b) {
- /* if (!a) {
-    a = 'null';
-  }
-  if (!b) {
-    b = 'null';
-  }*/
+const isEqual = function(a, b) {
   return a === b;
 };
 
-var sortBy = function(array = [], property = '', direction = 'asc') {
+const sortBy = function(array = [], property = '', direction = 'asc') {
   let arg;
   array.sort(function(a, b) {
     if (!a || !b || !a[property] || !b[property]) {
@@ -61,20 +55,46 @@ var sortBy = function(array = [], property = '', direction = 'asc') {
   });
 }
 
-var isString = function(string) {
+const sortDirs = (dirs) => {
+  dirs.sort(function(a, b) {
+    let prefCats = ['administration', 'preferences'];
+    let menuIdA = a.get_menu_id().toLowerCase();
+    let menuIdB = b.get_menu_id().toLowerCase();
+    let prefIdA = prefCats.indexOf(menuIdA);
+    let prefIdB = prefCats.indexOf(menuIdB);
+    if (prefIdA < 0 && prefIdB >= 0) {
+      return -1;
+    }
+    if (prefIdA >= 0 && prefIdB < 0) {
+      return 1;
+    }
+    let nameA = a.get_name().toLowerCase();
+    let nameB = b.get_name().toLowerCase();
+    if (nameA > nameB) {
+      return 1;
+    }
+    if (nameA < nameB) {
+      return -1;
+    }
+    return 0;
+  });
+  return dirs;
+};
+
+const isString = function(string) {
   return typeof string === 'string' || string instanceof String;
 }
 
-var unref = function(object) {
+const unref = function(object) {
   let keys = Object.keys(object);
-  for (var i = 0; i < keys.length; i++) {
+  for (let i = 0; i < keys.length; i++) {
     if (keys[i] !== 'buttonState' && keys[i] !== 'state') {
       object[keys[i]] = null;
     }
   }
 };
 
-var tryFn = function(fn, errCb) {
+const tryFn = function(fn, errCb) {
   try {
     return fn();
   } catch (e) {
@@ -83,3 +103,37 @@ var tryFn = function(fn, errCb) {
     }
   }
 };
+
+const map = function (arr, fn) {
+  if (arr == null) {
+    return [];
+  }
+
+  let len = arr.length;
+  let out = Array(len);
+
+  for (let i = 0; i < len; i++) {
+    out[i] = fn(arr[i], i, arr);
+  }
+
+  return out;
+}
+
+const find = function(arr, cb) {
+  for (let i = 0, len = arr.length; i < len; i++) {
+    if (cb(arr[i], i, arr)) {
+      return arr[i];
+    }
+  }
+  return null;
+}
+
+// TODO: Use this instead of queryCollection
+const findIndex = function(arr, cb) {
+  for (let i = 0, len = arr.length; i < len; i++) {
+    if (cb(arr[i], i, arr)) {
+      return i;
+    }
+  }
+  return -1;
+}